Renames package repositories to service
