+++ pom.xml
@@ -15,7 +15,7 @@
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
 		<version>1.4.3.RELEASE</version>
-		<relativePath/> <!-- lookup parent from repositories -->
+		<relativePath/> <!-- lookup parent from service -->
 	</parent>
 
 	<properties>
