--- src/main/java/com/exchange/services/GoodService.java
@@ -1,10 +1,0 @@
-package com.exchange.services;
-
-import org.springframework.stereotype.Service;
-
-/**
- * Created by optimize on 2/7/17.
- */
-@Service
-public class GoodService {
-}
